#207 add inline yaml path interface
