+++ src/netspresso_trainer/__init__.py
@@ -1,10 +1,12 @@
 from pathlib import Path
 
-from .cfg import TrainerConfig
-from .trainer_cli import parse_args_netspresso, set_arguments, train_cli
-from .trainer_inline import export_config_as_yaml, train_with_config
+from netspresso_trainer.cfg import TrainerConfig
+from netspresso_trainer.trainer_cli import parse_args_netspresso, set_arguments, train_cli
+from netspresso_trainer.trainer_inline import export_config_as_yaml, train_with_config, train_with_yaml
 
-train = train_with_config
+### Starting from v0.0.9, the default train function runs with yaml configuration
+train = train_with_yaml
+# train = train_with_config
 
 version = (Path(__file__).parent / "VERSION").read_text().strip()
